refactor: 実績の自動登録処理とユニットテストの整理
