+++ src/main/services/IActualAutoRegistrationFinalizerService.ts
@@ -1,0 +1,5 @@
+import { EventEntry } from '@shared/data/EventEntry';
+
+export interface IActualAutoRegistrationFinalizerService {
+  finalizeRegistration(mergedEvents: EventEntry[]): Promise<void>;
+}
