+++ src/main/services/IActualBuilderService.ts
@@ -1,0 +1,5 @@
+import { EventEntry } from '@shared/data/EventEntry';
+
+export interface IActualBuilderService {
+  buildActual(start: Date, end: Date): Promise<EventEntry | null>;
+}
