+++ src/main/services/__tests__/__mocks__/ActualAutoRegistrationFinalizerServiceMockBuilder.ts
@@ -1,0 +1,16 @@
+import { jest } from '@jest/globals';
+import { IActualAutoRegistrationFinalizerService } from '@main/services/IActualAutoRegistrationFinalizerService';
+import { EventEntry } from '@shared/data/EventEntry';
+
+export class ActualAutoRegistrationFinalizerServiceMockBuilder {
+  private finalizeRegistration: jest.MockedFunction<
+    (mergedActuals: EventEntry[]) => Promise<void>
+  > = jest.fn();
+
+  build(): IActualAutoRegistrationFinalizerService {
+    const mock: IActualAutoRegistrationFinalizerService = {
+      finalizeRegistration: this.finalizeRegistration,
+    };
+    return mock;
+  }
+}
