+++ src/main/types.ts
@@ -16,7 +16,6 @@ export const TYPES = {
   LabelService: Symbol.for('LabelService'),
   ProjectService: Symbol.for('ProjectService'),
   TaskService: Symbol.for('TaskService'),
-  ApplicationService: Symbol.for('ApplicationService'),
   PatternService: Symbol.for('PatternService'),
   GitHubCredentialsStoreService: Symbol.for('GitHubCredentialsStoreService'),
   UserPreferenceStoreService: Symbol.for('UserPreferenceStoreService'),
@@ -27,8 +26,10 @@ export const TYPES = {
   ActivityUsageService: Symbol.for('ActivityUsageService'),
   IpcService: Symbol.for('IpcService'),
   SpeakTextGenerator: Symbol.for('SpeakTextGenerator'),
-  AutoRegisterActualService: Symbol.for('AutoRegisterActualService'),
+  ActualAutoRegistrationService: Symbol.for('ActualAutoRegistrationService'),
+  ActualBuilderService: Symbol.for('ActualBuilderService'),
   OverlapEventMergeService: Symbol.for('OverlapEventMergeService'),
+  ActualAutoRegistrationFinalizerService: Symbol.for('ActualAutoRegistrationFinalizerService'),
 
   TaskScheduler: Symbol.for('TaskScheduler'),
   CalendarSyncProcessor: Symbol.for('CalendarSyncProcessor'),
