+++ src/renderer/src/services/IActualAutoRegistrationProxy.ts
@@ -1,7 +1,7 @@
 import { EventEntry } from '@shared/data/EventEntry';
 
-export interface IAutoRegisterActualService {
-  autoRegisterProvisionalActuals(targetDate: Date): Promise<EventEntry[]>;
+export interface IActualAutoRegistrationProxy {
+  autoRegisterProvisonalActuals(targetDate: Date): Promise<EventEntry[]>;
   confirmActualRegistration(targetDate: Date): Promise<void>;
   deleteProvisionalActuals(targetDate: Date): Promise<void>;
 }
