--- src/renderer/src/services/IAutoRegisterActualProxy.ts
@@ -1,7 +1,0 @@
-import { EventEntry } from '@shared/data/EventEntry';
-
-export interface IAutoRegisterActualProxy {
-  autoRegisterProvisonalActuals(targetDate: Date): Promise<EventEntry[]>;
-  confirmActualRegistration(targetDate: Date): Promise<void>;
-  deleteProvisionalActuals(targetDate: Date): Promise<void>;
-}
