+++ src/renderer/src/types.ts
@@ -12,14 +12,13 @@ export const TYPES = {
   ActivityEventProxy: Symbol.for('ActivityEventProxy'),
   GitHubEventProxy: Symbol.for('GitHubEventProxy'),
   ActicityUsageProxy: Symbol.for('ActivityUsageProxy'),
-  AutoRegisterActualProxy: Symbol.for('AutoRegisterActualProxy'),
+  ActualAutoRegistrationProxy: Symbol.for('ActualAutoRegistrationProxy'),
   CalendarSynchronizerProxy: Symbol.for('CalendarSynchronizerProxy'),
   GitHubSynchronizerProxy: Symbol.for('GitHubSynchronizerProxy'),
   CategoryProxy: Symbol.for('CategoryProxy'),
   LabelProxy: Symbol.for('LabelProxy'),
   ProjectProxy: Symbol.for('ProjectProxy'),
   TaskProxy: Symbol.for('TaskProxy'),
-  ApplicationProxy: Symbol.for('ApplicationProxy'),
   PatternProxy: Symbol.for('PatternProxy'),
 
   // service
