--- src/shared/data/Application.ts
@@ -1,10 +1,0 @@
-export interface Application {
-  id: string;
-
-  basename: string;
-  relatedProjectId: string;
-  relatedCategoryId: string;
-  relatedLabelIds: string[];
-
-  updated: Date;
-}
